// backend/src/core/training/enums.ts
export enum TRAINING_RUN_STATUS {
  NOT_STARTED = "not_started",
  IN_PROGRESS = "in_progress",
  COMPLETED = "completed",
}

export enum TRAINING_RUN_TASK_STATUS {
  NOT_STARTED = "not_started",
  IN_PROGRESS = "in_progress",
  PASSED = "passed",
  PASSED_WITH_NOTES = "passed_with_notes",
  FAILED = "failed",
}